remove outdated repatch, also install make rule
repatch was intended to avoid needing to update "changed" files on OS
upgrade, but the only relevant file is /etc/ttys, which very rarely
changes anyway, and even then hasn't broken compat in decades.

install was just a call to the script anyway and is superseded by the
zfs-specific scripts which automate away more of the workflow.